refactor: test suits and parser
