+++ c-compiler.cabal
@@ -30,10 +30,12 @@ library
   exposed-modules:
       Lib
       Compiler.Types
+      Compiler.Errors
       Compiler.Lexer.Parse
       Compiler.Lexer.Items
       Compiler.Lexer.LanguageDefinition
       Compiler.Parser.Grammar
+      Compiler.Parser.Common
       Compiler.Parser.Checkers
       Compiler.Generator.MASM
       Compiler.Generator.Emiters
