+++ src/Lib.hs
@@ -5,16 +5,13 @@ module Lib
 
 import           Compiler.Generator.MASM (generateFile, generateString)
 import           Compiler.Lexer.Parse    (parseFile, parseString)
-import           Compiler.Parser.Grammar (checkGrammar)
+import           Compiler.Parser.Grammar (checkFile, checkString)
+import           Compiler.Types          (ErrT)
 
 compileFile :: FilePath -> FilePath -> IO ()
-compileFile source destination =
-      parseFile source
-  >>= checkGrammar
-  >>= generateFile destination
+compileFile source destination = parseFile source >>=
+                                 checkFile >>=
+                                 generateFile destination
 
-compileString :: String -> IO String
-compileString source =
-      parseString source
-  >>= checkGrammar
-  >>= generateString
+compileString :: String -> Either ErrT String
+compileString = generateString . checkString . parseString
